Move vic-oil-gas-fields data source into its own file
The oil and gas sector implementation is likely to use a number of data sources, this will help keep them a bit better organised than all existing in a single file.
